Add two pages: package type and package owner

+++ database/migrations/2018_03_01_070530_create_package_types_table.php
@@ -22,7 +22,7 @@ class CreatePackageTypesTable extends Migration
             $table->integer('optional_package')->default(PackageType::NORMAL_PACKAGE);
             $table->integer('start_weight')->nullable();
             $table->integer('end_weight')->nullable();
-            $table->float('price');
+            $table->double('price');
             $table->timestamps();
         });
     }
